chore(frontend): temporary move of frontend components into frontend app
