+++ apps/frontend/pokemon-collection/components/card/CardList.tsx
