+++ apps/frontend/pokemon-collection/components/card/card.tsx
