+++ apps/frontend/pokemon-collection/components/common/FavoriteButton.tsx
