+++ apps/frontend/pokemon-collection/components/common/SharedMessageBar.tsx
