+++ apps/frontend/pokemon-collection/components/directory/RomDirectoryAside.tsx
