+++ apps/frontend/pokemon-collection/components/form/handleChange.ts
