+++ apps/frontend/pokemon-collection/components/form/handleClick.ts
