+++ apps/frontend/pokemon-collection/hooks/useFavorites.ts
