+++ apps/frontend/pokemon-collection/hooks/useSearchFilter.ts
