+++ apps/frontend/pokemon-collection/hooks/useSharedPageState.ts
