+++ apps/frontend/pokemon-collection/layouts/Footer.tsx
