+++ apps/frontend/pokemon-collection/layouts/Navbar.tsx
