+++ apps/frontend/pokemon-collection/layouts/header/SearchBar.tsx
