+++ apps/frontend/pokemon-collection/pages/Favorites.tsx
