+++ apps/frontend/pokemon-collection/pages/ProgressTrackerPage.tsx
