+++ apps/frontend/pokemon-collection/public/avalon.jpg
