+++ apps/frontend/pokemon-collection/public/golden-shield.png
