+++ apps/frontend/pokemon-collection/public/infinity.gif
