+++ apps/frontend/pokemon-collection/public/project-nova.jpg
