+++ apps/frontend/pokemon-collection/public/realidea-system.png
