+++ apps/frontend/pokemon-collection/public/stellar.png
