+++ apps/frontend/pokemon-collection/utils/itemListDrag.ts
